Relates to issue 586 : Checking whether relayLogDir is defined before trying to remove it in reset method
git-svn-id: http://tungsten-replicator.googlecode.com/svn/trunk@1350 9f5e3340-5c9a-25cb-7c1c-c133a2e9e928

+++ replicator/src/java/com/continuent/tungsten/replicator/management/ReplicationServiceManager.java
@@ -421,14 +421,18 @@ public class ReplicationServiceManager
         // Remove relay logs, if present.
         String relayLogDirName = serviceProps
                 .getString("replicator.extractor.dbms.relayLogDir");
-        File relayLogDir = new File(relayLogDirName);
-
-        if (!removeDirectory(relayLogDir, progress))
+        if (relayLogDirName != null)
         {
-            logger.error(String.format(
-                    "Could not remove the relay log directory %s", logDirName));
-        }
+            File relayLogDir = new File(relayLogDirName);
 
+            if (!removeDirectory(relayLogDir, progress))
+            {
+                logger.error(String.format(
+                        "Could not remove the relay log directory %s",
+                        logDirName));
+            }
+        }
+        
         logger.info("\n" + CLUtils.formatMap("progress", progress, "", false)
                 + "\n");
         return progress;
